LayoutTests/imported/w3c/web-platform-tests/custom-elements/scoped-registry/ShadowRoot-innerHTML.tentative-expected.txt


FAIL innerHTML on a shadow root should use scoped custom element registry shadowRoot.createElement is not a function. (In 'shadowRoot.createElement('some-element')', 'shadowRoot.createElement' is undefined)
FAIL innerHTML on a disconnected element should use the associated scoped custom element registry shadowRoot.createElement is not a function. (In 'shadowRoot.createElement('some-element')', 'shadowRoot.createElement' is undefined)
FAIL innerHTML on a connected shadow root should use its scoped custom element assert_true: expected true got false
FAIL innerHTML on an inserted element should use the scoped custom element of the ancestor shadow root shadowRoot1.createElement is not a function. (In 'shadowRoot1.createElement('some-element')', 'shadowRoot1.createElement' is undefined)
FAIL innerHTML should not upgrade a custom element inside a template element assert_equals: expected "SomeElement" but got "SomeElement1"
